requirements.txt
click
psycopg2-binary
python-dotenv